+++ CHANGELOG.md
@@ -21,7 +21,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - changed baidu map to mapbox for longitude and latitude
 - chenged microgrid list alarm link to fdd fault
 ### Fixed
-- added decimal data value range check to myems-modbus-tcp.
+- added decimal data value range check to myems-modbus-tcp
+- added missing lazy load files for router in myems-admin
 ### Removed
 
 ## [v4.3.0] - 2024-03-11